feat: add artifact context and run context utilities for AI tools
